requirements.txt
absl-py==1.0.0
addict==2.4.0
anyio==3.5.0
asgiref==3.5.0
basicsr==1.3.3.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.75.0
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h11==0.13.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
lmdb==1.3.0
Markdown==3.3.6
networkx==2.7.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pymongo==4.0.2
pyparsing==3.0.7
python-dotenv==0.19.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tb-nightly==2.9.0a20220316
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2022.3.16
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.0.3
yapf==0.32.0
zipp==3.7.0
